Fixed leasetime in IPv6 dhcp-range parameter of FTL/dnsmasq
Previously, when DHCP_LEASETIME was not set or set to '24h', then
FTL/dnsmasq failed with an error:

    dnsmasq: prefix length must be at least 64 at line 16 of
             /etc/dnsmasq.d/02-pihole-dhcp.conf

In those cases advanced/Scripts/webpage.sh set leasetime to "24" which
resulted in the following line in /etc/dnsmasq.d/02-pihole-dhcp.conf:

    dhcp-range=::100,::1ff,constructor:eth0,ra-names,slaac,24

But 24 is an integer and below 128 and thus FTL/dnsmasq interpreted it
as a IPv6 prefix length (cf. line 3139 in src/dnsmasq/option.c) which
was not intended.

This patch ensures that ${leasetime} is a string (due to the 'h') in all
if-then-else cases and thus correctly interpreted by FTL/dnsmasq. It also
ensures that an integer is passed to change_setting() so that
setupVars.conf is correct as well (i.e. no 'h' in setupVars.conf).

Fixes issue #4142